Draft implementation of definition primitive arguments support
